Makes mock initialization reusable by extracting it to TestControlDesigner.Mocks
